refactor :: review detail argument 수정

+++ feature/company/src/main/java/team/retum/company/navigation/CompanyDetailsNavigation.kt
@@ -12,7 +12,7 @@ const val NAVIGATION_COMPANY_DETAILS = "companyDetails"
 
 fun NavGraphBuilder.companyDetails(
     onBackPressed: () -> Unit,
-    navigateToReviewDetails: () -> Unit,
+    navigateToReviewDetails: (Long) -> Unit,
     navigateToReviews: (Long, String) -> Unit,
     navigateToRecruitmentDetails: (Long, Boolean) -> Unit,
 ) {
